+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 
 ## 1.1.0 - 2017-04-28
 
+- Add OpenStreetMap hillshade provider
 - Add Bing Roads provider
+- Add Short Names to providers
 
 ## 1.0.0 - 2017-04-01